### 2021-07-23 #3533 @whedon
[PRE REVIEW]: Lidium: a 3D point cloud processing software for TLS LiDAR data
**Submitting author:** @nsviel (<a href="http://orcid.org/0000-0002-4608-5850">Nathan</a>)
**Repository:** <a href="https://github.com/nsviel/Lidium" target ="_blank">https://github.com/nsviel/Lidium</a>
**Version:** v1.0.0
