### 2021-07-23 #3533 @danielskatz
👋 @nsviel - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). I also suspect there is something incorrect in your .bib file, since the check references command isn't returning. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` after making changes to the .md file or when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
