+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article{sanchiz_2021,
+@article{sanchiz:2021,
   TITLE = {{Radiometric Correction of Laser Scanning Intensity Data Applied For Terrestrial Laser Scanning.}},
   AUTHOR = {Sanchiz-Viel, Nathan and Bretagne, Estelle and Mouaddib, El Mustapha and Dassonvalle, Pascal},
   URL = {https://hal.archives-ouvertes.fr/hal-03042545},
@@ -12,7 +12,7 @@
   HAL_VERSION = {v1},
 }
 
-@Article{Kashani_2015,
+@Article{Kashani:2015,
 AUTHOR = {Kashani, Alireza G. and Olsen, Michael J. and Parrish, Christopher E. and Wilson, Nicholas},
 TITLE = {A Review of LIDAR Radiometric Processing: From Ad Hoc Intensity Correction to Rigorous Radiometric Calibration},
 JOURNAL = {Sensors},
@@ -27,14 +27,14 @@ ABSTRACT = {In addition to precise 3D coordinates, most light detection and rang
 DOI = {10.3390/s151128099}
 }
 
-@MISC{Eigen_2010,
+@MISC{Eigen:2010,
 author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
 title = {Eigen v3},
 howpublished = {http://eigen.tuxfamily.org},
 year = {2010}
 }
 
-@InProceedings{Pcl_2011,
+@InProceedings{Pcl:2011,
 author    = {Radu Bogdan Rusu and Steve Cousins},
 title     = {{3D is here: Point Cloud Library (PCL)}},
 booktitle = {{IEEE International Conference on Robotics and Automation (ICRA)}},
@@ -43,7 +43,7 @@ year      = {2011},
 address   = {Shanghai, China}
 }
 
-@inproceedings{flann_2009,
+@inproceedings{flann:2009,
 title={Flann, fast library for approximate nearest neighbors},
 author={Muja, Marius and Lowe, David G},
 booktitle={International Conference on Computer Vision Theory and Applications (VISAPP’09)},
@@ -52,7 +52,7 @@ year={2009},
 organization={INSTICC Press}
 }
 
-@Misc{Gnuplot_2013,                    
+@Misc{Gnuplot:2013,                    
 author={ Thomas Williams and Colin Kelley and {many others} },                
 title={ Gnuplot 4.6: an interactive plotting program },                    
 month={ April },                    
