+++ paper/paper.md
@@ -13,7 +13,7 @@ affiliations:
    index: 1
 date: 8 June 2021
 bibliography: paper.bib
-
+---
 
 # Summary
 
@@ -26,13 +26,13 @@ LiDAR (Light Detection and Ranging) sensors give huge amount of data, which gene
 
 `Lidium` is designed as a laboratory software to be used by both LiDAR researchers and by geomatic students for the manipulation and algorithm prototyping applied to point cloud data. This software is distinguished of the other open source solutions as CloudCompare (http://www.cloudcompare.org/) or Meshlab (https://www.meshlab.net/) by providing a lightweight and hackable framework to LiDAR data visualization and processing, with a focus on the point cloud registration step and the radiometric correction of the intensity information. This software aim to make 3D digitalization a way more accessible, as an alternative solution from proprietary software.
 
-The main difficulty of the three-dimensional reconstruction from the acquired point clouds, is the registration step. This one consists in the identification of the proper geometrical transformations allowing to regroup the point clouds in a common coordinate system. To do this, it is necessary to identify correspondences between the common areas of the clouds. This difficult problem concentrates the efforts of the research community. We use here an additional information acquired by the LiDAR, the intensity, as a discriminant element. This information is, by nature, insensible at external illuminations and related to the reflectance of the scanned material. However, the intensity is not widely used practically. Some effects induced by measure geometrical parameters and internal scanner treatments, make it strongly dependent to the viewpoint of the measure [@Kashani_2015]. 
+The main difficulty of the three-dimensional reconstruction from the acquired point clouds, is the registration step. This one consists in the identification of the proper geometrical transformations allowing to regroup the point clouds in a common coordinate system. To do this, it is necessary to identify correspondences between the common areas of the clouds. This difficult problem concentrates the efforts of the research community. We use here an additional information acquired by the LiDAR, the intensity, as a discriminant element. This information is, by nature, insensible at external illuminations and related to the reflectance of the scanned material. However, the intensity is not widely used practically. Some effects induced by measure geometrical parameters and internal scanner treatments, make it strongly dependent to the viewpoint of the measure [@Kashani:2015]. 
 
-We propose several approaches of radiometric correction and calibration, which under certain conditions, allow to make the intensity independent from the viewpoint and to convert it in a linear scale. A corrected intensity could also be useful in a large field of application : autonomous car for obstacle detection, clustering/segmentation, material or structure detection and recognition. 'Lidium' incorporate some research algorithms to proceed with the radiometric correction and calibration of this intensity. However, the intensity is scanner-dependent and the dataset used for the correction have to be made with the LiDAR of the user to get advantage of the present algorithms. The methodology and some algorithms are described in @sanchiz_2021.
+We propose several approaches of radiometric correction and calibration, which under certain conditions, allow to make the intensity independent from the viewpoint and to convert it in a linear scale. A corrected intensity could also be useful in a large field of application : autonomous car for obstacle detection, clustering/segmentation, material or structure detection and recognition. 'Lidium' incorporate some research algorithms to proceed with the radiometric correction and calibration of this intensity. However, the intensity is scanner-dependent and the dataset used for the correction have to be made with the LiDAR of the user to get advantage of the present algorithms. The methodology and some algorithms are described in [@sanchiz:2021].
 
[…]
