+++ build/CMakeFiles/TargetDirectories.txt
@@ -1,3 +1,3 @@
-/home/ns/Desktop/Lidium/build/CMakeFiles/rebuild_cache.dir
-/home/ns/Desktop/Lidium/build/CMakeFiles/edit_cache.dir
-/home/ns/Desktop/Lidium/build/CMakeFiles/executable.dir
+/home/aither/Desktop/Lidium/build/CMakeFiles/rebuild_cache.dir
+/home/aither/Desktop/Lidium/build/CMakeFiles/edit_cache.dir
+/home/aither/Desktop/Lidium/build/CMakeFiles/executable.dir

[…]
