+++ paper/paper.bib
@@ -10,6 +10,7 @@
   MONTH = Feb,
   HAL_ID = {hal-03042545},
   HAL_VERSION = {v1},
+  DOI = {https://doi.org/10.1016/j.isprsjprs.2020.11.015}
 }
 
 @Article{Kashani:2015,
