+++ paper/paper.bib
@@ -10,7 +10,7 @@
   MONTH = Feb,
   HAL_ID = {hal-03042545},
   HAL_VERSION = {v1},
-  DOI = {https://doi.org/10.1016/j.isprsjprs.2020.11.015}
+  DOI = {10.1016/j.isprsjprs.2020.11.015}
 }
 
 @Article{Kashani:2015,
@@ -42,6 +42,7 @@ booktitle = {{IEEE International Conference on Robotics and Automation (ICRA)}},
 month     = {May 9-13},
 year      = {2011},
 address   = {Shanghai, China}
+DOI       = {10.1109/ICRA.2011.5980567}
 }
 
 @inproceedings{flann:2009,
