+++ paper/paper.bib
@@ -10,7 +10,7 @@
   MONTH = Feb,
   HAL_ID = {hal-03042545},
   HAL_VERSION = {v1},
-  DOI = {10.1016/j.isprsjprs.2020.11.015}
+  DOI = {10.1016/j.isprsjprs.2020.11.015},
 }
 
 @Article{Kashani:2015,
@@ -25,7 +25,7 @@ URL = {https://www.mdpi.com/1424-8220/15/11/28099},
 PubMedID = {26561813},
 ISSN = {1424-8220},
 ABSTRACT = {In addition to precise 3D coordinates, most light detection and ranging (LIDAR) systems also record “intensity”, loosely defined as the strength of the backscattered echo for each measured point. To date, LIDAR intensity data have proven beneficial in a wide range of applications because they are related to surface parameters, such as reflectance. While numerous procedures have been introduced in the scientific literature, and even commercial software, to enhance the utility of intensity data through a variety of “normalization”, “correction”, or “calibration” techniques, the current situation is complicated by a lack of standardization, as well as confusing, inconsistent use of terminology. In this paper, we first provide an overview of basic principles of LIDAR intensity measurements and applications utilizing intensity information from terrestrial, airborne topographic, and airborne bathymetric LIDAR. Next, we review effective parameters on intensity measurements, basic theory, and current intensity processing methods. We define terminology adopted from the most commonly-used conventions based on a review of current literature. Finally, we identify topics in need of further research. Ultimately, the presented information helps lay the foundation for future standards and specifications for LIDAR radiometric calibration.},
-DOI = {10.3390/s151128099}
+DOI = {10.3390/s151128099},
 }
 
 @MISC{Eigen:2010,
@@ -41,8 +41,8 @@ title     = {{3D is here: Point Cloud Library (PCL)}},
 booktitle = {{IEEE International Conference on Robotics and Automation (ICRA)}},
 month     = {May 9-13},
 year      = {2011},
-address   = {Shanghai, China}
-DOI       = {10.1109/ICRA.2011.5980567}
+address   = {Shanghai, China},
+DOI       = {10.1109/ICRA.2011.5980567},
 }
 
 @inproceedings{flann:2009,
