### 2021-08-17 nsviel/Lidium#1 @hechth
JOSS Review checklist @hechth
## Review checklist for @hechth [openjournals/joss-reviews#3623](https://github.com/openjournals/joss-reviews/issues/3623)

- [ ] #2
