### 2021-09-13 #3623 @nsviel
Hi @martinfleis, unfortunately I don't have enough free time at the moment (as I start a new job). But I will try to resolved the opened questions as soon as possible. Some issues need a certain amount of work (automated tests, documentation), I could hope to finish at the end of September.
